+++ print-ntp.c
@@ -27,13 +27,13 @@
 #include "config.h"
 #endif
 
-#include <tcpdump-stdinc.h>
+#include <netdissect-stdinc.h>
 
 #ifdef HAVE_STRFTIME
 #include <time.h>
 #endif
 
-#include "interface.h"
+#include "netdissect.h"
 #include "addrtoname.h"
 #include "extract.h"
 
@@ -326,12 +326,12 @@ p_sfix(netdissect_options *ndo,
 {
 	register int i;
 	register int f;
-	register float ff;
+	register double ff;
 
 	i = EXTRACT_16BITS(&sfp->int_part);
 	f = EXTRACT_16BITS(&sfp->fraction);
-	ff = f / 65536.0;	/* shift radix point by 16 bits */
-	f = ff * 1000000.0;	/* Treat fraction as parts per million */
+	ff = f / 65536.0;		/* shift radix point by 16 bits */
+	f = (int)(ff * 1000000.0);	/* Treat fraction as parts per million */
 	ND_PRINT((ndo, "%d.%06d", i, f));
 }
 
@@ -344,15 +344,15 @@ p_ntp_time(netdissect_options *ndo,
 	register int32_t i;
 	register uint32_t uf;
 	register uint32_t f;
-	register float ff;
+	register double ff;
 
 	i = EXTRACT_32BITS(&lfp->int_part);
 	uf = EXTRACT_32BITS(&lfp->fraction);
 	ff = uf;
 	if (ff < 0.0)		/* some compilers are buggy */
 		ff += FMAXINT;
-	ff = ff / FMAXINT;	/* shift radix point by 32 bits */
-	f = ff * 1000000000.0;	/* treat fraction as parts per billion */
+	ff = ff / FMAXINT;			/* shift radix point by 32 bits */
+	f = (uint32_t)(ff * 1000000000.0);	/* treat fraction as parts per billion */
 	ND_PRINT((ndo, "%u.%09d", i, f));
 
 #ifdef HAVE_STRFTIME
@@ -381,7 +381,7 @@ p_ntp_delta(netdissect_options *ndo,
 	register uint32_t u, uf;
 	register uint32_t ou, ouf;
 	register uint32_t f;
-	register float ff;
+	register double ff;
 	int signbit;
 
 	u = EXTRACT_32BITS(&lfp->int_part);
@@ -419,8 +419,8 @@ p_ntp_delta(netdissect_options *ndo,
 	ff = f;
 	if (ff < 0.0)		/* some compilers are buggy */
 		ff += FMAXINT;
-	ff = ff / FMAXINT;	/* shift radix point by 32 bits */
-	f = ff * 1000000000.0;	/* treat fraction as parts per billion */
+	ff = ff / FMAXINT;			/* shift radix point by 32 bits */
+	f = (uint32_t)(ff * 1000000000.0);	/* treat fraction as parts per billion */
 	ND_PRINT((ndo, "%s%d.%09d", signbit ? "-" : "+", i, f));
 }
